# dist/api.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: 'Exercise 3.1'
  description: 'API - Web Development Environments'
...
paths:
  /Register:
    post:
      tags:
        - Auth
      summary: Add new user
      description: A new user from the client is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: User created on site
        "400":
          description: Not all reqired argument was given!
        "409":
          description: Username is already taken
        "401":
          description: User already logged in
  /Login:
    post:
      tags:
        - Auth
      summary: Login to the website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginScheme"
      responses:
        "200":
          description: >
            User was found in the DB and authenitcated successfully.
            The session ID is returned in a cookie named 'session'. 
            You should include this cookie in subsequent requests
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "400":
          description: Not all reqired argument was given!
        "401":
          description: UserName or Password is incorrect || User already login
...
  /search/{name}:
    get:
      tags:
      -  Search Pages
      summary: Search player, team and coach by player name 
      parameters: 
      - name: name 
        in: path
        required: true
        description: The name of the player / team / coach
        schema:
          type: string
          example: Rune 

      - name: teamName 
        in: query
        required: false
        description: Filter by team name
        schema:
          type: string  

      - name: position 
        in: query
        required: false
        description: Filter by position
        schema:
          type: number  
            
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
        '204':
          description: No items to show!
        '400':
          description: Bad request! 
  /search/{name}/sort/{sortParmeter}:
    get:
      tags:
      -  Search Pages
      summary: Search player, team and coach by player name 
      parameters: 
      - name: name 
        in: path
        description: The name of the player / team / coach
        required: true
        schema:
          type: string 

      - name: sortParmeter 
        in: path
        required: true
        description: The user can sort the result of the search by - playerName / teamName
        schema:
          type: string     
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Search'
        '204':
          description: No items to show!
        '400':
          description: SortBy values is not exists! 
...
  /manage_league/getAllGames:
    get:
      tags:
        - Association Representative
      summary: Return all the games in the system
      responses:
        "200":
          description: data about all the games in the system
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GamePreview"
        "401":
          description: User is not authorized
        "400":
          description: wrong input parameters
  /manage_league/addGame:
    post:
      tags:
        - Association Representative
      summary: Add new game
      description: A new game is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewGame"
      responses:
        "201":
          description: Game created on site
        "401":
          description: User is not authorized  
        "400":
          description: Not all reqired argument was given!
        "404":
          description: RefereeID is not exist || The game already accured  
  /manage_league/addResultToGame:
    put:
      tags:
      - Association Representative
      summary: Update result to a game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateGameResult'
      responses:
        '200':
          description: The result of the game where update!
        '401':
          description: User is not authorized 
        '400':
          description: Bad request!
        '404':
          description: The game has not yet been played          
  /manage_league/addListEventsToGame:
    post:
      tags:
      - Association Representative
      summary: Add list event to a game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateListEvent'
      responses:
        '201':
          description: The list of events has been successfully updated to the game
        '400':
          description: Bad request! 
        '401':
          description: User is not authorized 
        '404':
          description: The game has not yet been played     
  /manage_league/addReferee:
    post:
      tags:
      - Association Representative
      summary: Add new referee to our system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Referee'
      responses:
        '200':
          description: New referee added!
        '400':
          description: Bad request! 
        "401":
          description: User is not authorized   
  /manage_league/removeReferee/{refereeID}:
    delete:
      tags:
      -  Association Representative
      summary: Remove referee
      parameters: 
      - name: refereeID 
        in: path
        required: true
        schema:
          type: number 
      responses:
        '200':
          description: OK
        '204':
          description: No items to remove!
        '400':
          description: Bad request! 
        '401':
          description: User is not authorized  
        '404':
          description: The refereeID not found
...
components:
  schemas:
    User:
      type: object
      properties:      
        userName: 
          type: string
          minLength: 3
          maxLength: 8
          pattern: ^[a-zA-Z]+$
          description: The UserName must conatins 3 to 8 letters!
          example: Shahar

        firstName:
          type: string
          description: First name of the user
          example: Shahar
          
        lastName:
          type: string
          example: Tal
          description: Last name of ther user

        country:
          type: string
          description: Uses's country
          example: Israel
          
        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          pattern: ^[a-zA-Z0-9]*$
          example: Shahar123
          
        email:
          type: string 
          format: email
          example: ShaharAP@post.bgu.ac.il
          description: The email of the user
          
        image:
          type: string
          format: URL
          description: The user profile image
    LoginScheme:
      type: object
      properties:
        userName:
          type: string
          minLength: 3
          maxLength: 8
          description: The UserName must conatins 3 to 8 letters!
          example: ShaharAP

        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          description:  The Password must contains at least 5 to 8 characters, one letter and one digit
          example: Shahar123
    PlayerPreview:
      type: object
      properties:
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust

        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png

        position:
          type: integer
          description: player's position in team
          example: 1

        team_name:
          type: string
          description: player's team name
          example: København
    CoachPreview:
      type: object
      properties:
        firsName:
          type: string
          description: The coach first name
        
        lastName:
          type: string
          description: The coach last name
          
        team_name:
          type: string
          description: The name of the team where the coach coaches
          example: Barcelona
        
        image:
          type: string
          format: URL
          description: The coach image
...
    TeamPreviewSearch:
      type: object
      properties:
        name:
          type: string
          description: team name
          example: København

        logo_path:
          type: string
          description: logo_path
    GamePreview:
      type: object
      properties:
        game_date:
          type: string
          description: Game date
          example: '2021-06-01'

        game_time:
          type: string
          description: Game time
          example: '16:30'

        home_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Home team
          
        away_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Away team  

        referee_name:
          items: 
            $ref: '#/components/schemas/Referee'
          description: Referee Name

        field:
          type: string 
          description: The field of the game        
    NewGame:
      type: object
      properties:
        game_date:
          type: string
          description: game date
          example: "2021-06-01"

        game_time:
          type: string
          description: game date
          example: "16:30"   

        refereeID:
          type: number
          description: Thre RefereeID
          example: 3

        home_team_ID:
          type: number
          description: Home Team ID
          example: 85
          
        away_team_ID:
          type: number
          description: Home Team ID
          example: 939

        field:
          type: string 
          description: The field of the game
          example: Bloomfiled        
...
    Coach:
      type: object
      properties:
        name:
          type: string
          description: The coach first name
        
        team_name:
          type: string
          description: The name of the team where the coach coaches
          example: Barcelona
        
        image:
          type: string
          format: URL
          description: The coach image

        common_name:
          type: string
          description: player's common name
          example: Sten  

        nationality:
          type: string
          description: player's nation
          example: Israel

        birthdate:
          type: string
          description: player's birthday
          example: Israel

        birthcountry:
          type: string
          description: player's country of birth
          example: Israel  
    eventType: 
      type: string 
      enum:
      - Goal
      - Offsife
      - Offense
      - Red Card
      - Yellow card
      - Injury
      - Subsitution
    AddEvent:
      type: object
      properties:
        event_time:
          type: string
          description: The time of the event
          example: '16:30'
        
        minute:
          type: number
          description: The minute where the event accured
          example: 20
          
        event_description:
          type: string 
          description: Description of the event
          example: first goal of the game
          
        type:
          $ref: '#/components/schemas/eventType'
          example: 'Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'  
    Event:
      type: object
      properties:
        event_date:
          type: string
          description: The time of the event
          example: '01-01-2021'

        event_time:
          type: string
          description: The time of the event
          example: '16:30'
        
        minute:
          type: number
          description: The minute where the event accured
          example: 20
          
        event_description:
          type: string 
          description: Description of the event
          example: first goal of the game
          
        type:
          $ref: '#/components/schemas/eventType'
          example: 'Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'
    Game:
      type: object
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - field
      properties:          
        game_time:
          type: string
          description: game date

        home_result:
          type: number
          description: The result of the game for the home team
        
        away_result:
          type: number
          description: The result of the game for the away team  

        home_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Home team
          
        away_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Away team  
          
        referee_name:
          items: 
            $ref: '#/components/schemas/Referee'
          description: Referee Name  

        field:
          type: string 
          description: The field of the game
          
        event_list:
          type: array
          minLength: 3 
          items:
            $ref: '#/components/schemas/Event' 
    Team:
      type: object
      properties:
        name:
          type: string
          description: team name
          example: København

        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"

        future_games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        
        past_games:
          type: array
          items:
            $ref: '#/components/schemas/Game'    
            
        coach:
          type: object
          properties:
            items:
             $ref: '#/components/schemas/Coach'     
...
    updateGameResult:  
      type: object
      properties:
        gameID:
          type: number
          description: The game ID
          example: 1
        
        home_result:
          type: number
          description: The home team score
          example: 1  
      
        away_result:
          type: number
          description: The away team score
          example: 2
    updateListEvent:   
      type: object
      properties:
        gameID:
          type: number
          description: The game ID
          example: 1
          
        listEvents:
          type: array
          items:
            $ref: '#/components/schemas/AddEvent'
          description: 'Event type should be one of: Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'  
...
    Search:
      type: object
      properties:
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerPreview'
          
        teames:
          type: array
          items:
            $ref: '#/components/schemas/TeamPreviewSearch'

        coaches:
          type: array
          items:
            $ref: '#/components/schemas/CoachPreview'  
...
    Referee:
      type: object
      properties:
        user:
          description: Referee details
          items:
            $ref: '#/components/schemas/User'

        training:
          type: string
          description: The training of the referee
          example: Superliga trainer
...